Move token definitions to new file
