+++ src/parser.rs
@@ -1,3 +1,4 @@
+use crate::token::{Token, TokenType};
 use crate::{
     compiler::{Compiler, RollbackPoint, Span},
     errors::{Severity, SourceError},
@@ -198,70 +199,6 @@ impl AstNode {
             _ => 0,
         }
     }
-}
-
-#[derive(Debug)]
-pub enum TokenType {
-    Number,
-    Comma,
-    String,
-    Dollar,
-    Dot,
-    DotDot,
-    Name,
-    Pipe,
-    PipePipe,
-    Colon,
-    ColonColon,
-    Semicolon,
-    Plus,
-    PlusPlus,
-    PlusEquals,
-    Dash,
-    DashEquals,
-    Exclamation,
-    Asterisk,
-    AsteriskAsterisk,
-    AsteriskEquals,
-    ForwardSlash,
-    ForwardSlashForwardSlash,
-    ForwardSlashEquals,
-    Equals,
-    EqualsEquals,
-    EqualsTilde,
-    ExclamationTilde,
-    ExclamationEquals,
-    LParen,
-    LSquare,
-    LCurly,
-    LessThan,
-    LessThanEqual,
-    RParen,
-    RSquare,
-    RCurly,
-    GreaterThan,
-    GreaterThanEqual,
-    Ampersand,
-    AmpersandAmpersand,
-    QuestionMark,
-    ThinArrow,
-    ThickArrow,
-    Newline,
-    ErrGreaterThanPipe,
-    OutErrGreaterThanPipe,
-    OutGreaterThan,
-    OutGreaterGreaterThan,
-    ErrGreaterThan,
-    ErrGreaterGreaterThan,
-    OutErrGreaterThan,
-    OutErrGreaterGreaterThan,
-}
-
-#[derive(Debug)]
-pub struct Token {
-    pub token_type: TokenType,
-    pub span_start: usize,
-    pub span_end: usize,
 }
 
 impl Parser {
